Various fixes after merge from master
